cut out the part to use Rnd #56
